Reorganized registeration and email validation routes into more appropriate files.
