+++ app/routes/_auth.js
@@ -3,18 +3,10 @@ const isAuthenticated = require("./middleware/isAuthenticated");
 
 const {
     login,
-    register,
-    authenticated,
-    validateEmail,
-    resendValidation
+    authenticated
 } = require("../controllers/auth");
 
 router.post( "/authenticated", isAuthenticated, authenticated );
-
-router.post( "/register", register );
-
-router.post( "/validate/resend", isAuthenticated, resendValidation );
-router.post( "/validate/:token", validateEmail );
 
 router.post( "/login", login );
