+++ app/routes/register.js
@@ -1,0 +1,9 @@
+const router = require("express").Router();
+
+const {
+    register
+} = require("../controllers/register");
+
+router.post( "/", register );
+
+module.exports = router;
